fix(dapi): factored out a function and type to their own file
refs. metriport/metriport#865
